pull out validation utils into sep file
